% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_metrics.R
\name{ARI}
\alias{ARI}
\title{ARI}
\usage{
ARI(etas, eta_hats, n)
}
\arguments{
\item{etas}{Vector of true change-points}

\item{eta_hats}{Vector of estimated change-points}

\item{n}{Sample size}
}
\value{
The ARI
}
\description{
Computes the Adjusted Rand Index (ARI) of a vector of estimated change-points
}
\examples{
library(HDCD)
n = 400
true_changepoints = c(50,100)
est_changepoints = c(51,110)
ARI(true_changepoints, est_changepoints,n)
}
