% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_split_functions.R
\name{best_split_function_from_gain_function}
\alias{best_split_function_from_gain_function}
\title{Get a best_split_function from a gain_function}
\usage{
best_split_function_from_gain_function(gain_function, optimizer, control)
}
\arguments{
\item{gain_function}{A function with argiument \code{x}, \code{start}, \code{end}, \code{lambda}, \code{control} that returns a closure with arguments
split_point and possibly split_candidates that returns the gain after splitting the segment (\code{start}, \code{end}]
at \code{split_point} given data \code{x} and tuning parameter \code{lambda}.}

\item{optimizer}{Optimization method to be used to find an (approximate) of the maximum of the gain function.
One of \code{line_search} for BS or similar, \code{section_search} for OBS or similar or \code{two_step_search}.}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\value{
A function with formal arguments \code{x}, \code{start}, \code{end}, \code{split_candidates} and \code{lambda} that
 uses the optimizer specified to search for a maximum of the gain_function on the \code{split_candidates} given a segment
 (\code{start}, end], data \code{x} and a tuning parameter \code{lambda} and returns a list with arguments \code{gain}, \code{max_gain},
 \code{best_split} and possible \code{permutation_test} and \code{pval}.
}
\description{
Get a best_split_function from a gain_function
}
