% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{binary_segmentation}
\usage{
binary_segmentation(x, get_best_split, delta, lambda,
  model_selection_function = NULL, cross_validation_function = NULL,
  control = hdcd_control())
}
\arguments{
\item{x}{An n x p data matrix}

\item{get_best_split}{A function with formal arguments \code{x}, \code{start}, \code{end} and \code{split_candidates} that returns
a list with arguments \code{gain} and \code{best_split}, where gain is an array of length \code{nrow(x)} with possibly evaluations
of a gains curve or similar saved and \code{best_split} it the best element of \code{split_candidates} to split the interval 
(\code{start}, \code{end}].}

\item{delta}{The minimal relative segment length.}

\item{lambda}{A regularisation parameter for methods that require one.}

\item{model_selection_function}{A function with formal arguments \code{x}, \code{start}, \code{split_point} and \code{end} that
returns a list with arguments \code{statistic}, a value that measures the significance of the split at \code{split_point}, and
\code{is_siginificant}, a boolean indicating whether the value returned for \code{statistic} is significant.}

\item{cross_validation_function}{A function with formal arguments \code{x}, \code{start}, \code{end}, \code{lambda} and \code{folds} that returns
a list with arguments \code{cv_loss} and \code{lambda_opt}.}

\item{control}{An object of class \code{hdcd_control} as generated by \link{hdcd_control}.}
}
\description{
binary_segmentation
}
