% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_split_functions.R
\name{classifier_best_split_function}
\alias{classifier_best_split_function}
\title{Get a get_best_split function from a classifier}
\usage{
classifier_best_split_function(classifier, optimizer, control)
}
\arguments{
\item{classifier}{A function with formal arguments \code{x_train}, \code{y_train} and \code{control} that returns class probabilities in
a matrix of size nrow(x_train) x k, where k is the number of classes in y_train.}

\item{optimizer}{Optimization method to be used to find an (approximate) of the maximum of the gain function.
One of \code{line_search} for BS or similar, \code{section_search} for OBS or similar or \code{two_step_search}.}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\description{
Returns a closure with formal arguments \code{x}, \code{start}, \code{end} and \code{split_candidates} that finds the
maximizer of the gain_function given data \code{x}, \code{start} and \code{end} on \code{split_candidates}
using the optimizer \code{optimizer}.
}
