% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain_functions.R
\name{classifier_gain_function}
\alias{classifier_gain_function}
\title{Closure generating function to calculate gains for splits using a classifier}
\usage{
classifier_gain_function(x, start, end, classifier, control)
}
\arguments{
\item{x}{A matrix with observations in rows}

\item{start}{The start of the investigated interval}

\item{end}{The end of the investigated interval}

\item{classifier}{A function with formal arguments \code{x_train}, \code{y_train} and \code{control} that returns class probabilities in
a matrix of size nrow(x_train) x k, where k is the number of classes in y_train.}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\value{
A closure with parameters \code{x}, \code{start} and \code{end}, that when evaluated
will itself return a closure with arguments \code{split_point} and \code{split_candidates}.
Evaluated this function returns the gain when splitting the segment \code{(start, end]} of \code{x}
 at \code{split_point}.
}
\description{
Closure generating function to calculate gains for splits using a classifier
}
