% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{classifier_loglikelihood}
\alias{classifier_loglikelihood}
\title{Loglikelihood from classifier}
\usage{
classifier_loglikelihood(y_train, predictions, y_test = y_train, oob = F)
}
\arguments{
\item{y_train}{training labels}

\item{predictions}{matrix of predicted class probabilities}

\item{y_test}{test labels}

\item{oob}{boolean whether predictions are oob}
}
\description{
Loglikelihood from classifier
}
