% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{CompareChangepointsRand}
\alias{CompareChangepointsRand}
\title{CompareChangepointsRand}
\usage{
CompareChangepointsRand(cpts_a, cpts_b, n)
}
\arguments{
\item{cpts_a}{A sequence of changepoints.}

\item{cpts_b}{A sequence of changepoints.}

\item{n}{Total size of dataset from which both changepoint estimates originate.}
}
\value{
Returns a vector of the index values.
}
\description{
Rand type performance indices
}
\details{
Calculate rand type performance indices for two sets of changepoints. Typically one
of them will be the oracle estimate. See clues package for more details.
}
\examples{
CompareChangepointsRand(c(20, 50), c(30, 70), 100)
}
