% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{create_model}
\alias{create_model}
\title{Create GGM with changepoints}
\usage{
create_model(n, p, changepoints, model_function, mean_vecs = NULL, ...)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of dimensions}

\item{changepoints}{An array with the changepoints for the model to be generated. The change points are the last
observation of each segment.}

\item{model_function}{A function that spawns covariance matrices of dimension p.}

\item{mean_vecs}{If NULL, the mean for each segment will be zero.
Otherwise mean_vecs should be a list containing a p-dimensional numeric vector with the means for each segment.}

\item{...}{Addtional arguments to be supplied to modelFUN}
}
\value{
An object to be used by \link{simulate_from_model}
}
\description{
Create a model to generate data from a GGM for simulating the detection of changepoints
}
