% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{CreateModel}
\alias{CreateModel}
\title{CreateModel}
\usage{
CreateModel(n_segments, n, p, modelFUN, equispaced = T, mean_zero = T, ...)
}
\arguments{
\item{n_segments}{Number of segments in the model. There will be one less changepoints than segments.}

\item{n}{Number of observations}

\item{p}{Number of dimensions}

\item{modelFUN}{A function that spawns covariance matrices of dimension p.}

\item{equispaced}{If TRUE, the segments will be of equal length (up to rounding) and hence the changepoints will be equispaced.
Otherwise the changepoints will be drawn randomly and the distance between them will differ.}

\item{mean_zero}{If TRUE, the mean for each segment will be zero.
Otherwise it will be sampled from a standard normal distribution for each segment.}

\item{...}{Addtional arguments to be supplied to modelFUN}
}
\value{
An object to be used by \link{SimulateFromModel}
}
\description{
Create a model to generate data from for simulating the detection of changepoints
}
