% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DiagMatrice}
\alias{DiagMatrice}
\title{DiagMatrice}
\usage{
DiagMatrice(p)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}
}
\value{
A covariance matrix.
}
\description{
Spawn a p-dimensional identity matrix.
}
\examples{
DiagMatrice(50)
}
