% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{DiagMatrix}
\alias{DiagMatrix}
\title{DiagMatrix}
\usage{
DiagMatrix(p)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}
}
\value{
A covariance matrix.
}
\description{
Spawn a p-dimensional identity matrix.
}
\examples{
DiagMatrix(50)
}
