% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_segments}
\alias{draw_segments}
\title{Draw segment (-boundaries) for WBS or BS}
\usage{
draw_segments(n, n_segments = n, delta = 0.1, segmentation = "BS", alpha = 1/2)
}
\arguments{
\item{n}{total number of observations available}

\item{n_segments}{Number of segments to be drawn (only relevant of WBS)}

\item{delta}{Minimal relative segment length}

\item{segmentation}{One of BS, WBS or SBS}

\item{alpha}{decay parameter in [1/2, 1) for SBS}
}
\description{
Draw segment (-boundaries) for WBS or BS
}
