% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FullRegression}
\alias{FullRegression}
\title{FullRegression}
\usage{
FullRegression(x, cpts, lambda = 0.1, standardize = T, threshold = 1e-07)
}
\arguments{
\item{x}{A n times p data matrix}

\item{cpts}{A set of changepoints correspoding to row indices.}

\item{lambda}{Sparsity penalty parameter}
}
\value{
\describe{
  \item{est_coefs}{A p times p matrix of estimated coefficients.}
  \item{est_mean}{A p vector of estimated means.}
  \item{est_intercepts}{A p vector of estimated intercepts.}
}
}
\description{
Neighbourhood selection for a given set of changepoints and a data sequence
}
