% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_split_function_from_gain_function.R
\name{get_best_split_function_from_gain_function}
\alias{get_best_split_function_from_gain_function}
\title{Get a best split function from a gain function}
\usage{
get_best_split_function_from_gain_function(gain_function, optimizer,
  control)
}
\arguments{
\item{gain_function}{A function with formal arguments \code{x}, \code{start}, \code{end} and \code{lambda} that returns
a closure with argument \code{split_point}, that returns the gain after splitting the segment (\code{start}, \code{end}]
at \code{split_point} given data \code{x} and tuning parameter \code{lambda}.}
}
\value{
A function with formal arguments \code{x}, \code{start}, \code{end}, \code{split_candidates} and \code{lambda} that
 uses the optimizer specified to search for a maximum of the gain_function on the \code{split_candidates} given a segment
 (\code{start}, end], data \code{x} and a tuning parameter \code{lambda} and returns a list with arguments \code{gain}, an
 array of length end - start with evaluations of the gain function and \code{best_split}.
}
\description{
Returns a closure with formal arguments \code{x}, \code{start}, \code{end} and \code{split_candidates} that finds the
maximizer of the \code{gain_function} given data \code{x}, \code{start} and \code{end} on \code{split_candidates}
using the optimizer \code{optimizer}.
}
