% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_change_points_from_tree}
\alias{get_change_points_from_tree}
\title{Get Change Points from a binary_segmentation_tree}
\usage{
get_change_points_from_tree(tree, variable = "cv_improvement",
  value = 0)
}
\arguments{
\item{tree}{An object of class \strong{binary_segmentation_tree}}

\item{variable}{Name of the variable with respect to which should be pruned. Usually one of \code{cv_improvement} or
\code{max_gain}.}

\item{value}{Value until which the tree will be pruned}
}
\value{
A vector with the sorted change points.
}
\description{
Utility function to get the change points with high enough value for some variable from a binary_segmentation_tree
}
