% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_util.R
\name{get_cov_mat}
\alias{get_cov_mat}
\title{Calculate a covariance matrix}
\usage{
get_cov_mat(x, NA_method = c("complete_observations",
  "pairwise_covariance", "loh_wainwright_bias_correction",
  "average_imputation"), min_points = 2)
}
\arguments{
\item{x}{A n x p design matrix}

\item{NA_method}{The method that should be applied to estimate the covariance structure.}

\item{min_points}{Minimal number of available observations of a variable necessary such that an estimation of the variance of that 
variable is reliable. An integer not smaller than 2.}
}
\value{
A p x p matrix with an estimate for the covariance structure of \code{x}. If less than \code{min_points} observations are
available for some variable, the corresponding rows and columns are NA.
}
\description{
Calculate a covariance matrix of the design matrix \code{x} using one of the methods 
\code{complete_observations}, \code{pairwise_covariance}, \code{loh_wainwright_bias_correction} or \code{average_imputation}.
}
