% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_util.R
\name{get_glasso_fit}
\alias{get_glasso_fit}
\title{Get a glasso fit}
\usage{
get_glasso_fit(x, lambda, control)
}
\arguments{
\item{x}{A design matrix}

\item{lambda}{Regularisation parameter for the glasso function.}

\item{control}{An object of class \code{hdcd_control} generated by \link{hdcd_control}.}
}
\value{
A list with values \code{cov_mat}, the original estimate of the covariance structure,
\code{inds} an array indicating the variables for which a variance could be estimated, 
the precision matrix estimated with the glasso \code{wi} and its inverse \code{w}.
}
\description{
Get a glasso fit
}
