% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain_functions.R
\name{get_glasso_gain_function}
\alias{get_glasso_gain_function}
\title{Closure generating function to calculate gains when splitting and learning a glasso model}
\usage{
get_glasso_gain_function(control)
}
\arguments{
\item{control}{an object of type \code{hdcd_control} returned by \link{hdcd_control}}
}
\value{
A closure with parameters \code{x}, \code{start} and \code{end}, that when evaluated
will itself return a closure with parameter \code{split_point}. This calculates the gain when
splitting the segment \code{(start, end]} of \code{x} at \code{split_point}.
}
\description{
Closure generating function to calculate gains when splitting and learning a glasso model
}
