% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_splits}
\alias{get_splits}
\title{obtain all possible segmentation scenarios that can be obtained from the input x by pruning}
\usage{
get_splits(x)
}
\arguments{
\item{x}{tree as generated by \link{binary_segmentation}}
}
\description{
obtain all possible segmentation scenarios that can be obtained from the input x by pruning
}
