% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{GetChangePointsFromLeafs}
\alias{GetChangePointsFromLeafs}
\title{GetChangePointsFromLeafs}
\usage{
GetChangePointsFromLeafs(x)
}
\arguments{
\item{x}{An object of class \strong{bs_tree}}
}
\value{
A vector with the sorted changepoints.
}
\description{
Utility function to get the changepoints from the name of the leaf nodes in the tree.
}
