% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{glasso_cross_validation_function}
\alias{glasso_cross_validation_function}
\title{Glasso cross validation function}
\usage{
glasso_cross_validation_function(x, start, end, lambda, folds, control)
}
\arguments{
\item{x}{A matrix with observations in rows}

\item{start}{The start of the investigated interval}

\item{end}{The end of the investigated interval}

\item{lambda}{Tuning parameter passed to the method.}

\item{folds}{folds}

\item{control}{an object of type \code{hdcd_control} returned by \link{hdcd_control}}
}
\description{
Glasso cross validation function
}
