% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HandleGlassoNaN}
\alias{HandleGlassoNaN}
\title{HandleGlassoNaN}
\usage{
HandleGlassoNaN(w)
}
\arguments{
\item{w}{Warning to be handled}
}
\description{
Suppress specific warning in glasso package
}
\details{
To be used with withCallingHandlers. Glasso produces non-pd covariance matrices
when using the Meinshausen-Bühlmann approximation. It then tries to calculate the
likelihood an hence log(det(cov_mat_i)) which will throw a warning if cov_mat is
not pd.
}
