% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{hdcd_control}
\alias{hdcd_control}
\title{Create an object of class hdcd_control to supply parameters to \link{hdcd}}
\usage{
hdcd_control(
  glasso_cv_inner = TRUE,
  glasso_cv_inner_lambda = NULL,
  glasso_cv_inner_randomize_folds = FALSE,
  glasso_cv_inner_n_folds = 5,
  glasso_cv_inner_lambda_stepsize = sqrt(2),
  glasso_NA_method = "loh_wainwright_bias_correction",
  glasso_standardize = TRUE,
  glasso_penalize_diagonal = FALSE,
  classifier_oob = TRUE,
  section_search_min_points = 5,
  section_search_stepsize = 0.1,
  random_forest_n_tree = 600,
  random_forest_mtry = NULL,
  random_forest_sample_fraction = 1,
  kNN_k = function(x) ceiling(sqrt(x)),
  wbs_n_segments = 100,
  sbs_alpha = 1/sqrt(2),
  permutation_test = FALSE,
  permutation_test_pvalue = 0.05,
  permutation_test_n = 400
)
}
\arguments{
\item{glasso_cv_inner}{Should change points be selected according to inner cross validation in the glassocd procedure.}

\item{glasso_cv_inner_lambda}{Grid of values for lambda to be analyzed during inner cross validation. If NULL, values will
be searched for iteratively}

\item{glasso_cv_inner_randomize_folds}{Should inner folds be selected in a randomized fashion (instead of equispaced)}

\item{glasso_cv_inner_n_folds}{Number of inner folds}

\item{glasso_NA_method}{Method to be used for covariance estimation in the presence of missing values. One of
\code{loh_wainwright_bias_correction}, \code{pairwise} or \code{average}. See accompanying paper for more details.}

\item{glasso_standardize}{Should values be be standardized  before they are fed into the glasso?}

\item{glasso_penalize_diagonal}{Should values on the diagonal of the precision matrix be penalized in the glasso?}

\item{classifier_oob}{Does the (custom) classifier return oob predictions?}

\item{section_search_stepsize}{Stepsize parameter for the section_search optimizer}

\item{random_forest_n_tree}{n_tree parameter for Random Forest algorithm ranger}

\item{random_forest_mtry}{mtry parameter for Random Forest algorithm ranger}

\item{random_forest_sample_fraction}{sample_fraction parameter for Random Forest algorithm ranger}

\item{kNN_k}{Number of nearest neighbors used for classification. Either a positive integer or a function
that reuturns a positive integer given the number of total observations.}

\item{wbs_n_segments}{Number of segments to be drawn for the WBS procedure}

\item{sbs_alpha}{Decay parameter for the SBS procedure}

\item{permutation_test}{should a permutation test be done (approximated) for model selection (only relevant for kNNcd and RFcd)}

\item{permutation_test_pvalue}{pvalue threshold used in permutation test}

\item{permutation_test_n}{number of permutations (approximations) in permutation test}
}
\description{
Create an object of class hdcd_control to supply parameters to \link{hdcd}
}
