% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{HubNetwork}
\alias{HubNetwork}
\title{HubNetwork}
\usage{
HubNetwork(p, max_hubs = p/10)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}

\item{n_hubs}{Number of hubs in the network}
}
\value{
A covariance matrix.
}
\description{
Spawn a hub network covariance matrix
}
\examples{
HubNetwork(50)
}
