% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{InitNaiveSquaredLoss}
\alias{InitNaiveSquaredLoss}
\title{Square loss with naive calculation}
\usage{
InitNaiveSquaredLoss(x)
}
\arguments{
\item{x}{A n times p matrix or data frame.}
}
\description{
Square loss with naive calculation
}
