% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_split_functions.R
\name{kNN_best_split_function}
\alias{kNN_best_split_function}
\title{Get a get_best_split function from the kNN classifier}
\usage{
kNN_best_split_function(x, control)
}
\arguments{
\item{x}{A matrix with observations in rows}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\description{
Returns a closure with formal arguments \code{x}, \code{start}, \code{end} and \code{split_candidates} that finds the
maximizer of the gain_function given data \code{x}, \code{start} and \code{end} on \code{split_candidates}
using the optimizer \code{optimizer}.
}
