% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain_functions.R
\name{kNN_gain_function}
\alias{kNN_gain_function}
\title{kNN gain function}
\usage{
kNN_gain_function(x, control = hdcd_control())
}
\arguments{
\item{x}{A matrix with observations in rows}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\description{
kNN gain function
}
