% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{line_search}
\alias{line_search}
\title{Line search optimisation algorithm}
\usage{
line_search(gain_function, split_candidates, n, control)
}
\arguments{
\item{gain_function}{closure}

\item{split_candidates}{An array of indices where gain function is (possibly) evaluated.}

\item{n}{total number of observations}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\value{
A list with entries gain, max_gain and best_split
}
\description{
Finds an (exact) maximum of \code{gain_function} by evaluating it at each of \code{split_candidates}.
}
