% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_eps}
\alias{log_eps}
\title{Bounded approximation to the logarithm}
\usage{
log_eps(x, threshold = exp(-6))
}
\arguments{
\item{x}{numeric}

\item{threshold}{bound}
}
\description{
Bounded approximation to the logarithm
}
