% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_space}
\alias{log_space}
\title{Logarithmically Scaled Sequence Generation}
\usage{
log_space(from, to, length.out)
}
\arguments{
\item{from}{the starting value of the sequence}

\item{to}{the end value of the sequence}

\item{length.out}{the length of the sequence}
}
\description{
Generates a logarithmically scaled sequence
}
