% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_util.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Negative loglikelihood of a multivariate normal}
\usage{
loglikelihood(x, mu, cov_mat_inv)
}
\arguments{
\item{x}{A design matrix}

\item{mu}{Mean of the multivariate Gaussian distribution}

\item{cov_mat_inv}{precision matrix of the multivariate Gaussian distribution}
}
\value{
An array with the negative log-likelihoods of the observations in \code{x} given the Gaussian model
specified.
}
\description{
Negative loglikelihood of a multivariate normal
}
