% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{MoveEdges}
\alias{MoveEdges}
\title{MoveEdges}
\usage{
MoveEdges(prec_mat, share_moves = 0.1)
}
\arguments{
\item{share_moves}{Share of edges to be moved.}

\item{mat}{A precision matrix}
}
\description{
Radomly move a share of the edges in a random graph
}
\details{
In order to create a slightly different graph with the same level of sparsity
the selected share of edges will be randomly moved to positions where no edge existed
before. Make sure to choose share_moves low for dense graphs.
}
