% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{PenalizeSplitsFUN}
\alias{PenalizeSplitsFUN}
\title{PenalizeSplitsFUN}
\usage{
PenalizeSplitsFUN(gamma)
}
\arguments{
\item{gamma}{Split penality}
}
\value{
FALSE if the node should be pruned from tree
}
\description{
Utility function to determine whether a node should be pruned from tree for
given gamma value. This is a closure and hence returns a function parametrized with gamma
to be used as prune function.
}
