% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{plot.binary_segmentation_tree}
\alias{plot.binary_segmentation_tree}
\title{S3 Method for plotting an object of class binary_segmentation_tree}
\usage{
\method{plot}{binary_segmentation_tree}(x, true_change_points = NULL,
  ...)
}
\arguments{
\item{x}{An object of class binary_segmentation_tree}

\item{true_change_points}{an array containing the true underlying change points}

\item{...}{Further arguments passed to print generic.}
}
\description{
Simulatenously plots all gain curves of a binary_segmentation_tree
}
