% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{plot.bs_cv}
\alias{plot.bs_cv}
\title{plot.bs_cv}
\usage{
\method{plot}{bs_cv}(x, ..., show_legend = T)
}
\arguments{
\item{x}{An object of class \strong{bs_cv}}

\item{...}{Only included to be consist with plot generic.}

\item{show_legend}{If TRUE the legend is shown on the bottom of the plot.}
}
\description{
S3 method for plotting the results of cross-validation.
}
