% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_missingness_structure}
\alias{plot_missingness_structure}
\title{Plot the missingness structure of a design matrix}
\usage{
plot_missingness_structure(x, order = F)
}
\arguments{
\item{x}{design matrix}

\item{order}{Should the variables of x be reoderd to better display blocks?}
}
\description{
Plot the missingness structure of a design matrix
}
