% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{PostProcess}
\alias{PostProcess}
\title{PostProcess}
\usage{
PostProcess(x, cpts, delta, lambda, method)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{cpts}{A numeric vector containing the found changepoints. Can be of length zero if no changepoints have been found.}

\item{delta}{Value between 0 and 0.5. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{lambda}{Sparsity penality parameter in single lasso fits.}

\item{method}{Which method should be used for fitting the model? See defaults for possible choices.}
}
\value{
A nuemric vector containing the final segment boundaries
}
\description{
Adjust the position of each changepoint by refitting the model in the respective segments.
}
