% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{print.binary_segmentation_tree}
\alias{print.binary_segmentation_tree}
\title{print.binary_segmentation_tree}
\usage{
\method{print}{binary_segmentation_tree}(x, ...)
}
\arguments{
\item{x}{A data.tree node.}

\item{...}{Further arguments passed to print generic.}
}
\description{
S3 method for printing a binary_segmentation_tree object.
}
\details{
Decorate the print method of the data.tree package to see more details at each node.
}
