% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcd.R
\name{print.bs_cv}
\alias{print.bs_cv}
\title{Print method for objects of class bs_cv}
\usage{
\method{print}{bs_cv}(x, ...)
}
\arguments{
\item{x}{Object of class bs_cv}

\item{...}{Only included to be consist with print generic.}
}
\description{
Print method for objects of class bs_cv
}
