% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{RandomNetwork}
\alias{RandomNetwork}
\title{RandomNetwork}
\usage{
RandomNetwork(p, prob = 3/p)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}

\item{prob}{Probabilty that a pair of nodes have a common edge.}
}
\value{
A covariance matrix.
}
\description{
Spawn a erdos renyi type random network covariance matrix
}
\examples{
RandomNetwork(50)
}
