% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ScaleNetwork}
\alias{ScaleNetwork}
\title{ScaleNetwork}
\usage{
ScaleNetwork(p, preferential_power = 1, u = 0.1, v = 0.3,
  prec_mat = T, scaled = F)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}

\item{preferential_power}{Power coefficient alpha for weighting of degree number as alpha in prefential attachment mechanism.}

\item{u}{Constant added to the diagonal elements of the precision matrix for controlling the magnitude of partial correlations.}

\item{v}{Constant added to the off diagonal of the precision matrix for controlling the magnitude of partial correlations.}

\item{prec_mat}{Should the precision matrix be returned? If false the covariance matrix will be returned (default).}

\item{scaled}{If TRUE the created precision matrix will be inverted and scaled to a correlation matrix.}
}
\value{
A covariance or precision matrix.
}
\description{
Spawn a hub network covariance matrix
}
\examples{
ScaleNetwork(50)
}
