% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{section_search}
\alias{section_search}
\title{Section search optimisation algorithm}
\usage{
section_search(gain_function, split_candidates, n, control)
}
\arguments{
\item{gain_function}{closure}

\item{split_candidates}{An array of indices where gain function is (possibly) evaluated.}

\item{n}{total number of observations}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\value{
A list with entries gain, max_gain and best_split
}
\description{
Finds an (approximate) maximum of \code{gain_function} by intelligently evaluating at some of the \code{split_candidates}.
Works best if \code{split_candidates} is an interval.
}
