% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_in_mean_and_variance}
\alias{shift_in_mean_and_variance}
\title{find best split shift in mean and variance}
\usage{
shift_in_mean_and_variance(x, ...)
}
\arguments{
\item{x}{Array with entries that are assumed to have a shift in mean and variance at some split point.}

\item{alpha}{array of segment boundaries}

\item{train_fold}{array containing indices in training fold}
}
\value{
An array on length \code{length(x)} with gains resulting from splitting at that specific split point.

The negative gaussian loglikelihood of observations x for estimated mean and variance is \eqn{-n/2 * (log(2 \pi \hat\sigma^2) + 1)}.
The gaussian maximum likelihood estimate \eqn{\hat\sigma^2} is \code{(sum(x^2) - sum(x)^2/length(x))/length(x)}
}
\description{
Efficiently calculates the gain in a one-dimensional shift in mean and variance scenario.
}
