% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_from_model}
\alias{simulate_from_model}
\title{Simulate Observations from a model created by create_model}
\usage{
simulate_from_model(model)
}
\arguments{
\item{model}{An object as created by \link{create_model}}
}
\value{
A n times p matrix of simulated data.
}
\description{
Simulate Observations from a model created by create_model
}
\examples{

# Simulate 100 observations from a 10-dimensional chain network with two change points
model <- create_model(3, 100, 10, ChainNetwork)
x <- simulate_from_model(model)
}
