% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_non_parametric}
\alias{simulate_non_parametric}
\title{simulate (non-parametric) data sets}
\usage{
simulate_non_parametric(model, seed)
}
\arguments{
\item{model}{One if \code{Iris}, \code{Noise}, \code{Wine_1}, \code{Wine_2}, \code{RN},
\code{CIM} or \code{CIC}}

\item{seed}{seed used for dataset creation}
}
\description{
simulate (non-parametric) data sets
}
