% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimulateFromModel}
\alias{SimulateFromModel}
\title{SimulateFromModel}
\usage{
SimulateFromModel(model)
}
\arguments{
\item{model}{An object as created by \link{CreateModel}}
}
\value{
A n times p matrix of simulated data.
}
\description{
SimulateFromModel
}
\examples{

# Simulate 100 observations from a 10-dimensional chain network with two changepoints
mod <- CreateModel(3, 100, 10, ChainNetwork)
SimulateFromModel(mod)
}
