% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{TernarySearch}
\alias{TernarySearch}
\title{TernarySearch}
\usage{
TernarySearch(split_candidates, left, right, x, SegmentLossFUN, intervals = 3)
}
\arguments{
\item{split_candidates}{A list of indices of possible split points}

\item{left}{start index on the left}

\item{right}{start index on the right}

\item{x}{An n times p data matrix}

\item{SegmentLossFUN}{A loss function is created by \code{\link{SegmentLoss}}}

\item{intervals}{Number of intervals of the search grid}
}
\description{
Use a ternary search technique to find a local minimum for the next split recursively.
}
\details{
Besides the standard ternary approach where the segment is divided into three intervals adn evaluated at two points one
can also define a more granular intervals until intervals = number of observations in which case all observations will
be evaluated and we are in the case of searching for the minium exactly.
}
