% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{Divide data into training + testing data}
\usage{
train_test_split(x, alpha, train_fold = 1:nrow(x))
}
\description{
Divide data into training + testing data
}
