% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{Binary Segmentation}
\usage{
binary_segmentation(
  x,
  get_best_split,
  delta,
  lambda = 0,
  gamma = 0,
  segmentation = "BS",
  cross_validation_function = NULL,
  control = hdcd_control()
)
}
\arguments{
\item{x}{An n x p data matrix}

\item{get_best_split}{A function with arguments \code{x}, \code{start}, \code{end}, \code{split_candidates} and \code{lambda}, that returns a list with
entries gain, max_gain, best_split and possibly permutation_test and pval.}

\item{delta}{Minimal relative segment length, defaults to 0.1}

\item{lambda}{Tuning parameter passed to the method.}

\item{gamma}{Minimal required gain to split in the Binary Segmentation Procedure. Defaults to 0.}

\item{segmentation}{Segmentation method to be used. One of \code{BS}, \code{SBS} or \code{WBS}.}

\item{cross_validation_function}{A function with arguments \code{x}, \code{start}, \code{end}, \code{lambda}, \code{folds}, \code{control}}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\description{
Search for change points with Binary Segmentation, Wild Binary Segmentation or Seeded Binary Segmentation
}
