% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{BinarySegmentation}
\alias{BinarySegmentation}
\title{BinarySegmentation}
\usage{
BinarySegmentation(x, delta, lambda, method = c("nodewise_regression",
  "summed_regression", "ratio_regression"), penalize_diagonal = F,
  optimizer = c("line_search", "ternary_search", "section_search"),
  control = NULL, standardize = T, threshold = 1e-07, verbose = FALSE,
  ...)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{delta}{Value between 0 and 0.5. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{lambda}{Sparsity penality parameter in single lasso fits.}

\item{method}{Which method should be used for fitting the model? See defaults for possible choices.}

\item{penalize_diagonal}{Boolean, should the diagonal elements of the precision matrix be penalized?}

\item{control}{}

\item{threshold}{The threshold for halting the iteration in glasso or glmnet. In the former it controls the change of single parameters
in the latter it controls the total objective value.}
}
\value{
An object of class \strong{bs_tree}
}
\description{
Uses the binary segmentation algorithmn in order to build a binary tree of the data sequence
and given tuning parameters lambda and delta recursively. The tree can then be pruned in order to obtain
a changepoint estimate.
}
\examples{
dat <- SimulateFromModel(CreateModel(n_segments = 2,n = 100,p = 30, ChainNetwork))
res <- BinarySegmentation(dat, delta = 0.1, lambda = 0.01, method = "summed_regression")
print(res)
}
