% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ChainNetwork}
\alias{ChainNetwork}
\title{ChainNetwork}
\usage{
ChainNetwork(p, n_perm = p, a = 0.5, prec_mat = F)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}

\item{n_perm}{Positive integer. The first n_perm dimensions will be permuted randomly.}

\item{a}{Positive float between 0 and 1. Scale parameter for the elements of the covariance matrix}

\item{cov_mat}{Should the precision matrix be returned? If false the covariance matrix will be returned (default).}
}
\value{
A covariance matrix.
}
\description{
Spawn a chain network covariance matrix
}
\examples{
ChainNetwork(50)
}
