% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{CrossValidation}
\usage{
CrossValidation(x, delta, lambda = seq(0.01, 0.1, 0.01),
  method = c("summed_regression", "ratio_regression"), threshold = 1e-04,
  penalize_diagonal = F, use_ternary_search = F, n_folds = 10,
  gamma_max = NULL, verbose = T)
}
\arguments{
\item{x}{A n times p matrix for which to find the best splitting point.}

\item{delta}{Value between 0 and 1. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{lambda}{Sparsity penality parameter in single lasso fits.}

\item{method}{Which method should be used for fitting the model? See defaults for possible choices.}

\item{threshold}{The threshold for halting the iteration in glasso or glmnet. In the former it controls the change of single parameters
in the latter it controls the total objective value.}

\item{penalize_diagonal}{Boolean, should the diagonal elements of the precision matrix be penalized?}

\item{use_ternary_search}{Use a ternary search algorithm in each level of the recursion to find a local optimum (EXPERIMNETAL)}

\item{n_folds}{Number of folds. Test data will be selected equi-spaced, i.e. each n_fold-th observation.}

\item{gamma_max}{If NULL the range of gamma will be determined by doing an extra fit on the full model and taking the
difference between the full segment loss and the loss of the first split.}

\item{verbose}{If TRUE additional information will be printed.}
}
\value{
A nested list with the cv results and the full fitted models for each gamm, lambda combination.
}
\description{
EXPERIMENTAL. Crossvalidation for the desired method and parameter combinations.
Evalusate lambda values will refit the model whereas gamma values can be evaluted using a
single fit.
}
\details{
Will make use of a registered parallel backend over the folds. This will be changed to parallelize over lambdas
in each fold to make use of all nodes if the number of folds is smaller than the number of nodes.
}
\examples{
dat <- SimulateFromModel(CreateModel(n_segments = 2,n = 100,p = 30, ChainNetwork))
CrossValidation(dat, delta = 0.1, method = "summed_regression")
}
