% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{FindBestSplit}
\alias{FindBestSplit}
\title{FindBestSplit}
\usage{
FindBestSplit(x, delta, n_obs, use_ternary_search, SegmentLossFUN, intervals)
}
\arguments{
\item{x}{A n times p matrix for which to find the best splitting point.}

\item{delta}{Value between 0 and 1. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{use_ternary_search}{Use a ternary search algorithm in each level of the recursion to find a local optimum (EXPERIMNETAL)}

\item{SegmentLossFUN}{A loss function is created by \code{\link{SegmentLoss}}}

\item{intervals}{Number of intervals of the search grid}
}
\description{
Uses the segment_loss function for each possible split into two segements and
returns the best splitting index as well as the loss for each candidate.
}
