% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{FindBestSplit}
\alias{FindBestSplit}
\title{FindBestSplit}
\usage{
FindBestSplit(x, delta, n_obs, optimizer = c("line_search", "ternary_search",
  "section_search"), control, SegmentLossFUN)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{delta}{Numeric value between 0 and 0.5. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{n_obs}{The number of observations in the data set.}

\item{optimizer}{Which search technique should be used for performing individual splits in the binary segmentation alogrithm? Possible choices are
\itemize{
  \item line_search: Exhaustive linear search. All datapoints are evaluated and the maximum is returned.
  \item ternary_search: Iteratively cuts the search space according by a fixed ratio and approximately finds a local maximum.
  \item section_search: Iteratively cuts the search space according by a flexible ratio as determined by stepsize in control parameter and approximately finds a local maximum.
}}

\item{control}{A list with parameters that is accessed by the optimizer.
\itemize{
  \item stepsize: Numeric value between 0 and 0.5. Used by section search.
  \item intervals: Integer value larger than 3. Used by ternary search.
}}

\item{SegmentLossFUN}{A loss function as created by \code{\link{SegmentLoss}}}
}
\description{
Uses the SegmentLossFUN function for each possible split on the given segment
considering delta and returns the splitting index with the lowest loss
as well as the loss for each candidate.
}
