% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FullRegression}
\alias{FullRegression}
\title{FullRegression}
\usage{
FullRegression(x, cpts, lambda = 0.1, standardize = T, threshold = 1e-07)
}
\arguments{
\item{x}{A n times p data matrix}

\item{cpts}{A set of changepoints correspoding to row indices.}

\item{lambda}{Sparsity penalty parameter.}

\item{standardize}{Boolean. If TRUE the penalty parameter \eqn{\lambda} will
be adjusted for every dimension in the single Lasso fits according to the
standard deviation in the data.}

\item{threshold}{The threshold for halting the iteration in
\code{\link[glasso]{glasso}} or \code{\link[glmnet]{glmnet}}. In the former
it controls the absolute change of single parameters in the latter it
controls the total objective value. This value is ignored if FUN is not
NULL.}
}
\value{
\describe{
  \item{est_coefs}{A p times p matrix of estimated coefficients.}
  \item{est_mean}{A p vector of estimated means.}
  \item{est_intercepts}{A p vector of estimated intercepts.}
}
}
\description{
Neighbourhood selection for a given set of changepoints and a data sequence
}
