% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{get_change_points_from_tree}
\alias{get_change_points_from_tree}
\title{Get Change Points from a binary_segmentation_tree}
\usage{
get_change_points_from_tree(tree)
}
\arguments{
\item{tree}{An object of class \strong{binary_segmentation_tree}}
}
\value{
A vector with the sorted changepoints.
}
\description{
Utility function to get the change points with positive value for some variable from a binary_segmentation_tree
}
