% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcd.R
\name{hdcd}
\alias{hdcd}
\title{hdcd}
\usage{
hdcd(x, delta = 0.1, lambda = NULL, gamma = NULL,
  method = c("nodewise_regression", "summed_regression", "ratio_regression"),
  penalize_diagonal = F, optimizer = c("line_search", "ternary_search",
  "section_search"), control = NULL, standardize = T, threshold = 1e-07,
  n_folds = 10, verbose = F, parallel = T, ...)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{delta}{Value between 0 and 0.5. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{lambda}{Sparsity penality parameter in single lasso fits.}

\item{gamma}{Split penalty parameter for pruning the tree. If NULL k-fold cross-validation will be conducted.}

\item{method}{Which method should be used for fitting the model? See defaults for possible choices.}

\item{penalize_diagonal}{Boolean, should the diagonal elements of the precision matrix be penalized?}

\item{control}{}

\item{threshold}{The threshold for halting the iteration in glasso or glmnet. In the former it controls the change of single parameters
in the latter it controls the total objective value.}

\item{n_folds}{Number of folds in cross validation. Default is 10.}

\item{verbose}{Should additional information be printed?}

\item{...}{Supply additional arguments for a specific method (e.g. p for nodewise_regression)}
}
\value{
The indices of the found changepoints in the sequence
}
\description{
High Dimensional Changepoint Detection
}
\examples{
dat <- SimulateFromModel(CreateModel(n_segments = 2,n = 100,p = 30, ChainNetwork))
hdcd(dat, 0.1, 0.1, 0.05, method = "summed_regression", verbose = T)
}
