% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcd.R
\name{hdcd}
\alias{hdcd}
\title{High-dimensional change point detection}
\usage{
hdcd(
  x,
  delta = 0.1,
  gamma = 0,
  lambda = NULL,
  method = NULL,
  optimizer = "line_search",
  segmentation = "BS",
  control = hdcd_control()
)
}
\arguments{
\item{x}{A matrix with observations in rows}

\item{delta}{Minimal relative segment length, defaults to 0.1}

\item{gamma}{Minimal required gain to split in the Binary Segmentation Procedure. Defaults to 0.}

\item{lambda}{Tuning parameter passed to the method.}

\item{method}{Method to be applied for change point detection. One of \code{glassocd}, \code{RFcd}, \code{kNNcd},
\code{custom_classifier}, \code{custom_gain_function}, \code{custom_best_split_function}.}

\item{optimizer}{Optimization method to be used to find an (approximate) of the maximum of the gain function.
One of \code{line_search} for BS or similar, \code{section_search} for OBS or similar or \code{two_step_search}.}

\item{segmentation}{Segmentation method to be used. One of \code{BS}, \code{SBS} or \code{WBS}.}

\item{control}{Control parameter as returned by \link{hdcd_control}}
}
\description{
Find breaks in distribution in a possiblity high-dimensional time series
}
