% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{hdcd_control}
\alias{hdcd_control}
\title{Create an object of class hdcd_control to supply parameters to \link{hdcd}}
\usage{
hdcd_control(cv_inner = FALSE, cv_inner_lambda = NULL,
  cv_inner_randomize_folds = FALSE, cv_inner_n_lambda = 10,
  cv_inner_n_folds = 10, cv_inner_nrep_em = 5,
  cv_inner_search_lambda = F, cv_inner_min_grid_ratio = 0.01,
  cv_inner_lambda_step = (0.1)^(1/10), cv_inner_stop_early = FALSE,
  glasso_standardize = TRUE, glasso_penalize_diagonal = FALSE,
  glasso_threshold = 1e-04,
  glasso_NA_method = "loh_wainwright_bias_correction",
  nodewise_regression_node = 1, elastic_net_alpha = 1,
  elastic_net_family = "gaussian", verbose = F, max_depth = Inf,
  section_search_min_points = 5, section_search_stepsize = 0.5,
  section_search_k_sigma = 0, section_search_tolerance = 0,
  segment_loss_min_points = 2, n_obs = NULL)
}
\description{
Create an object of class hdcd_control to supply parameters to \link{hdcd}
}
