% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{line_search}
\alias{line_search}
\title{Line search optimisation algorithm}
\usage{
line_search(gain_function, split_candidates, control)
}
\arguments{
\item{gain_function}{A function with formal arguments \code{x}, \code{start}, \code{end} and \code{lambda} that returns
a closure with argument \code{split_point}, that returns the gain after splitting the segment (\code{start}, \code{end}]
at \code{split_point} given data \code{x} and tuning parameter \code{lambda}.}

\item{split_candidates}{array of indices where gain function is evaluated.}

\item{control}{An object of class \code{hdcd_control} as generated by \link{hdcd_control}.}
}
\value{
An array of length \code{control$n_obs} with entries for gains at all of \code{split_candidates}.
}
\description{
Finds an (exact) maximum of \code{gain_function} by evaluating it at each of \code{split_candidates}.
}
