% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{MoveEdges}
\alias{MoveEdges}
\title{MoveEdges}
\usage{
MoveEdges(x, share_moves = 0.1, tol = 1e-16)
}
\arguments{
\item{x}{A precision matrix}

\item{share_moves}{Share of edges to be moved.}

\item{tol}{Tolerance for zero values.}
}
\description{
Radomly move a share of the edges in a random graph
}
\details{
In order to create a slightly different graph with the same level of sparsity
the selected share of edges will be randomly moved to positions where no edge existed
before. Make sure to choose share_moves low for dense graphs.
}
