% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{plot.bs_cv}
\alias{plot.bs_cv}
\title{plot.bs_cv}
\usage{
\method{plot}{bs_cv}(results)
}
\arguments{
\item{cv_results}{An object of class \strong{bs_cv}}
}
\description{
S3 method for plotting the results of cross-validation.
}
