% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{PostProcess}
\alias{PostProcess}
\title{PostProcess}
\usage{
PostProcess(x, cpts, delta, lambda, method)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{cpts}{A numeric vector containing the found changepoints. Can be of length zero if no changepoints have been found.}

\item{delta}{Numeric value between 0 and 0.5. Tuning param which determines the minimal segment size
proportional to the size of the dataset and hence an upper bound for the number of changepoints.}

\item{lambda}{Positive numeric value. This is the regularization parameter in the single Lasso fits.}

\item{method}{Which estimator should be used? Possible choices are
\itemize{
  \item nodewise_regression: Nodewise regression is based on a single node that needs to be specified with an additional parameter p.
  \item summed_regression: Summed nodewise regression sums up the residual variances of nodewise regression over all nodes.
  \item ratio_regression: Likelihood ratio based regression sums the pseudo-profile-likelihood over all nodes.
}}
}
\value{
A numeric vector containing the final segment boundaries
}
\description{
Adjust the position of each changepoint by refitting the model in the respective segments.
}
