% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{RandomNetwork}
\alias{RandomNetwork}
\title{RandomNetwork}
\usage{
RandomNetwork(
  p,
  prob = min(1, 5/p),
  u = 0.1,
  v = 0.3,
  prec_mat = F,
  scaled = F
)
}
\arguments{
\item{p}{Positive integer.The desired number of dimensions.}

\item{prob}{Probabilty that a pair of nodes have a common edge.}

\item{u}{Constant added to the diagonal elements of the precision matrix for controlling the magnitude of partial correlations.}

\item{v}{Constant added to the off diagonal of the precision matrix for controlling the magnitude of partial correlations.}

\item{prec_mat}{Should the precision matrix be returned? If false the covariance matrix will be returned (default).}

\item{scaled}{If TRUE the created precision matrix will be inverted and scaled to a correlation matrix.}
}
\value{
A covariance matrix.
}
\description{
Spawn a erdos renyi type random network covariance matrix
}
\examples{
RandomNetwork(50)
}
