% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_folds}
\alias{sample_folds}
\title{Sample Folds}
\usage{
sample_folds(n, k, randomize = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of folds}

\item{randomize}{should folds be randomized? Equispaced folds are returned for randomize = FALSE}
}
\description{
Sample Folds
}
