% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{SectionSearch}
\alias{SectionSearch}
\title{SectionSearch}
\usage{
SectionSearch()
}
\value{
Returns a function with arguments split_candidates, left, mid, right, x, SegmentLossFUN, RecFUN, stepsize
where RecFun should always be set to the object name of the function has been assigned so the function can
call itself recursively.
}
\description{
Implements a variant of golden section search where the stepsize can be choosen
freely.
}
\details{
Hence its is possible to trade robustness against computational effciency. This function is
a closure that implements a cache for already calculated values and returns a function that can
be used to find the global minimum numerically.
}
