% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{SplitLoss}
\alias{SplitLoss}
\title{SplitLoss}
\usage{
SplitLoss(x, split_point, SegmentLossFUN, start, end)
}
\arguments{
\item{x}{A n times p data matrix.}

\item{split_point}{Index on which to split the segment}

\item{SegmentLossFUN}{A loss function as generate by closure \link{SegmentLoss}.}
}
\value{
Sum of the loss for both new segments after split.
}
\description{
Caluclates the sum of the loss that results from splitting the segment at the given point
}
