% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{SplitLoss}
\alias{SplitLoss}
\title{SplitLoss}
\usage{
SplitLoss(x, split_point, SegmentLossFUN)
}
\arguments{
\item{opt_split}{Index on which to split the segment}
}
\value{
Sum of the loss for both new segments after split.
}
\description{
Caluclates the sum of the loss that results from splitting the segment at the given point
}
